number.py (#9)
WB публикует metadata как 100.0, то есть строкой с десятичной точкой. int("100.0") вызывает ошибку:
Error creating entity for device wb-mao4_21: invalid literal for int() with base 10: '100.0'

+++ wirenboard/entities/number.py
@@ -13,17 +13,18 @@ logger = logging.getLogger(__name__)
 class WirenBoardNumber(WirenBoardEntity, NumberEntity):
     """Representation of a Wiren Board number."""
 
-    def __init__(self, device_info, mqtt_client):
+    def __init__(self, device_info, mqtt_client):      
         value = device_info.get("min", "0")
         if value:
-            self._attr_native_min_value = int(value)
+            self._attr_native_min_value = float(value)
         else:
-            self._attr_native_min_value = 0
+            self._attr_native_min_value = 0.0
+
         value = device_info.get("max", "100")
         if value:
-            self._attr_native_max_value = int(value)
+            self._attr_native_max_value = float(value)
         else:
-            self._attr_native_max_value = 100
+            self._attr_native_max_value = 100.0  
         if (
             device_info["device_id"] == "buzzer"
             and device_info["control_id"] == "frequency"
